% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ratios.R
\name{aggregate_ratios}
\alias{aggregate_ratios}
\title{Calculate Aggregate Ratios and Save as RDS File}
\usage{
aggregate_ratios(df, site_dir)
}
\arguments{
\item{df}{A data frame containing the variables `Date`, `mean_scattered`, `dir_diff_ratio`, `GPSLatitude`, `GPSLongitude`, and `Date2`.}

\item{site_dir}{The directory path where the RDS file should be saved.}
}
\value{
A data frame with aggregated values of mean_scattered, dir_diff_ratio, Lat_mean, Long_mean, and Date_mean.
}
\description{
This function calculates aggregate ratios such as mean scattered ratio and directional difference ratio.
The function also calculates the mean latitude and longitude, and mean Date for each unique Date in the data frame.
The results are saved as an RDS file in the specified directory.
}
\examples{
\dontrun{
  aggregate_ratios(df = your_df, site_dir = "your_directory_path_here")
}
}
