% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_and_save_sun_angle.R
\name{compute_and_save_sun_angle}
\alias{compute_and_save_sun_angle}
\title{Compute sun angle and save the results as an RDS file}
\usage{
compute_and_save_sun_angle(site_dir)
}
\arguments{
\item{site_dir}{The site directory where the input and output RDS files are located in /CSV/Corrected_values.}
}
\value{
None. This function saves the computed data as an RDS file.
}
\description{
This function reads an RDS file containing XMP data, computes the sun angle based on SolarElevation,
SolarAzimuth, Roll, Pitch, and Yaw, and then saves the result as an RDS file.
}
\examples{
\dontrun{
compute_and_save_sun_angle("your_directory_here")
}
}
