% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_corrected_irradiance.R
\name{compute_corrected_irradiance}
\alias{compute_corrected_irradiance}
\title{Compute Corrected Horizontal Irradiance}
\usage{
compute_corrected_irradiance(xmp_all_ssa, ratios)
}
\arguments{
\item{xmp_all_ssa}{A data frame containing the needed variables including `Date`, `BandName`, `Irradiance`, etc.}

\item{ratios}{A data frame containing the ratios needed for the correction.}
}
\value{
A data frame containing the corrected irradiance values.
}
\description{
This function computes the corrected horizontal irradiance for each image, considering the Fresnel
correction and the direct and scattered components of irradiance.
}
\examples{
\dontrun{
  corrected_irradiance_df <- compute_corrected_irradiance(xmp_all_ssa = your_data_frame_here, ratios = your_ratios_here)
}
}
