% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rolling_regression.R
\name{compute_rolling_regression}
\alias{compute_rolling_regression}
\title{Compute Rolling Regression and Subsequent Filtering}
\usage{
compute_rolling_regression(site_dir, xmp_all_ssa, lookback_seconds = 30)
}
\arguments{
\item{site_dir}{The directory path where the RDS file should be saved.}

\item{xmp_all_ssa}{The data frame containing the relevant variables.}

\item{lookback_seconds}{The lookback window in seconds for the rolling regression.}
}
\value{
A data frame with refined results of the rolling regression.
}
\description{
This function computes a rolling regression for the given data.
It fits a linear model to estimate the relationship between Irradiance and cos_SSA.
It then applies additional transformations and filtering to the data.
}
\examples{
\dontrun{
  result_df <- compute_rolling_regression(site_dir = "your_directory_path_here", xmp_all_ssa = your_data_frame_here)
}
}
