% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scatter_direct_ratio.R
\name{compute_scatter_direct_ratio}
\alias{compute_scatter_direct_ratio}
\title{Compute Scatter:Direct Ratio for Each Flight}
\usage{
compute_scatter_direct_ratio(
  site_dir,
  r_squared_threshold = 0.4,
  rolling_window = 30
)
}
\arguments{
\item{site_dir}{The directory where the input RDS file is located.}

\item{r_squared_threshold}{The R-squared threshold to use (default is 0.4).}

\item{rolling_window}{The time window for the rolling regression (default is 30 seconds).}
}
\value{
A data frame containing computed values and models.
}
\description{
This function performs several operations to estimate scattered:direct ratio for each flight.
}
\examples{
\dontrun{
compute_scatter_direct_ratio("your_directory_here", r_squared_threshold = 0.4, rolling_window = 30)
}
}
