% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sun_angle.R
\name{compute_sun_angle}
\alias{compute_sun_angle}
\title{Compute Sun-Sensor Angle}
\usage{
compute_sun_angle(SolarElevation, SolarAzimuth, Roll, Pitch, Yaw)
}
\arguments{
\item{SolarElevation}{Numeric. The elevation angle of the sun above the horizon.}

\item{SolarAzimuth}{Numeric. The azimuth angle of the sun.}

\item{Roll}{Numeric. The roll angle of the drone.}

\item{Pitch}{Numeric. The pitch angle of the drone.}

\item{Yaw}{Numeric. The yaw angle of the drone.}
}
\value{
Numeric. The angle between the sun and the sensor in radians.
}
\description{
This function calculates the angle between the sun and the Downwelling Light Sensor (DLS)
based on the sun's position and the orientation of the drone.
}
\examples{
angle = compute_sun_angle(40, 120, 0.2, 0.1, 0.3)
}
