% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresnel_correction.R
\name{fresnel_correction}
\alias{fresnel_correction}
\title{Compute Fresnel Correction for Downwelling Light Sensor (DLS)}
\usage{
fresnel_correction(x)
}
\arguments{
\item{x}{A data frame or a list containing at least the columns/elements `Irradiance` and `SunSensorAngle`.
`Irradiance` is the measured irradiance value, and `SunSensorAngle` is the angle between
the sun and the sensor, measured in radians.}
}
\value{
The Fresnel transmission coefficient, which is a correction factor that can be
        applied to the raw irradiance measurements.
}
\description{
This function calculates the Fresnel correction factor for irradiance measured by
a Downwelling Light Sensor (DLS). It uses the function `multilayer_transmission` to
apply Fresnel equations for multilayer transmission.
}
\examples{
fresnel_correction(data.frame(Irradiance = c(1000, 1100), SunSensorAngle = c(0.1, 0.2)))
}
