% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresnel_transmission.R
\name{fresnel_transmission}
\alias{fresnel_transmission}
\title{Fresnel Transmission for Downwelling Light Sensor (DLS)}
\usage{
fresnel_transmission(phi, n1 = 1, n2 = 1.5, polarization = c(0.5, 0.5))
}
\arguments{
\item{phi}{The angle of incidence in radians. It's the angle between the incident ray and the surface normal.}

\item{n1}{The refractive index of the first medium (typically air).}

\item{n2}{The refractive index of the second medium (typically the material of the DLS sensor).}

\item{polarization}{A numeric vector of length 2 containing the weights of s-polarization (Rs) and p-polarization (Rp).}
}
\value{
The Fresnel transmission coefficient T, clamped between 0 and 1.
}
\description{
This function computes the Fresnel transmission for the Downwelling Light Sensor (DLS)
using the Fresnel equations for reflection and transmission of an electromagnetic wave
when it crosses the boundary between two media with different refractive indices.
This correction adjusts for the DLS reflecting, rather than measuring, some of the irradiance that hits it.
The function is based on the equations from the MicaSense GitHub repository.
}
\examples{
fresnel_transmission(0.5, 1, 1.5, c(0.5, 0.5))
}
