% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilayer_transmission.R
\name{multilayer_transmission}
\alias{multilayer_transmission}
\title{Compute Multilayer Transmission using Fresnel Equations}
\usage{
multilayer_transmission(phi, n, polarization)
}
\arguments{
\item{phi}{The initial angle of incidence in radians. It's the angle between the incident ray and the surface normal of the first layer.}

\item{n}{A numeric vector representing the refractive indices of each layer in the multilayer material, starting with the outermost layer.}

\item{polarization}{A numeric vector of length 2 containing the weights of s-polarization (Rs) and p-polarization (Rp).}
}
\value{
The total Fresnel transmission coefficient T, calculated for the multilayer material.
}
\description{
This function calculates the transmission coefficient for a multilayer material
based on the Fresnel equations. It uses the function `fresnel_transmission` to calculate
the transmission for each layer and multiplies them together.
}
\examples{
multilayer_transmission(0.5, c(1, 1.5, 1.2), c(0.5, 0.5))
}
