% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geolocation.R
\name{plot_geolocation}
\alias{plot_geolocation}
\title{Plot Geolocation against Sun Sensor Angle}
\usage{
plot_geolocation(xmp_corrected, band_name = "NIR")
}
\arguments{
\item{xmp_corrected}{A data frame containing the corrected irradiance values.}

\item{band_name}{The band name to filter ("NIR" is default).}
}
\value{
A ggplot2 plot.
}
\description{
This function plots the GPSLongitude and GPSLatitude colored by SunSensorAngle.
}
\examples{
\dontrun{
  plot_geolocation(xmp_corrected, "NIR")
}
}
