% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_irradiance.R
\name{plot_irradiance}
\alias{plot_irradiance}
\title{Plot Sensor and Horizontal Irradiance}
\usage{
plot_irradiance(xmp_corrected, band_name = "Blue")
}
\arguments{
\item{xmp_corrected}{A data frame containing the corrected irradiance values.}

\item{band_name}{The band name to filter ("Blue" is default).}
}
\value{
A ggplot2 plot.
}
\description{
This function plots the Sensor and Horizontal Irradiance against Date2.
}
\examples{
\dontrun{
  plot_irradiance(xmp_corrected, "Blue")
}
}
