% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scattered_light.R
\name{plot_scattered_light}
\alias{plot_scattered_light}
\title{Generate Scattered Light Ratio Plot}
\usage{
plot_scattered_light(df)
}
\arguments{
\item{df}{A data frame containing the necessary variables, including `Date2`, `percent_scattered`, and `R2`.}
}
\value{
A ggplot object
}
\description{
This function generates a ggplot to visualize the rolling regression results. It plots the percent of scattered light over time.
The color of the points represents the R-squared values of the linear models used in the rolling regression.
}
\examples{
\dontrun{
  plot_scattered_light(df = your_df)
}
}
