% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_pattern.R
\name{plot_spatial_pattern}
\alias{plot_spatial_pattern}
\title{Plot Spatial Pattern of Sun Sensor Angles}
\usage{
plot_spatial_pattern(df)
}
\arguments{
\item{df}{A data frame containing the necessary variables, including `GPSLongitude`, `GPSLatitude`, `SunSensorAngle`, and `Date`.}
}
\value{
A ggplot object
}
\description{
This function generates a ggplot to visualize the spatial pattern of Sun Sensor Angles.
It's a scatter plot with the longitude and latitude as coordinates, and the sun sensor angle as the color.
}
\examples{
\dontrun{
  plot_spatial_pattern(df = your_df)
}
}
