% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sun_sensor_angles.R
\name{plot_sun_sensor_angles}
\alias{plot_sun_sensor_angles}
\title{Plot Sun-Sensor Angles}
\usage{
plot_sun_sensor_angles(df)
}
\arguments{
\item{df}{A data frame that contains the Date, BandName, Date2, and SunSensorAngle columns.}
}
\value{
A ggplot object.
}
\description{
This function takes a data frame containing the sun-sensor angle and plots it to
check if the angles are within a reasonable range.
}
\examples{
\dontrun{
plot_sun_sensor_angles(xmp_all_ssa)
}
}
