% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif_and_save.R
\name{read_exif_and_save}
\alias{read_exif_and_save}
\title{Read EXIF data and save it}
\usage{
read_exif_and_save(site_dir)
}
\arguments{
\item{site_dir}{The directory where the site's data is located.}
}
\value{
Nothing, but saves EXIF data to RDS files.
}
\description{
Read EXIF data and save it
}
\examples{
\dontrun{
read_exif_and_save(site_dir = "your_site_directory")
}
}
