% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds_and_combine.R
\name{read_rds_and_combine}
\alias{read_rds_and_combine}
\title{Combine RDS Files Into a Single DataFrame}
\usage{
read_rds_and_combine(site_dir)
}
\arguments{
\item{site_dir}{The directory where the image files and rds are located.}
}
\value{
A data frame containing all combined EXIF data.
}
\description{
This function reads in multiple RDS files based on a list of dates and band numbers,
and combines them into a single dataframe.
}
\examples{
\dontrun{
result <- read_rds_and_combine(dir = "your_directory", site_dir = "your_site_directory")
}
}
