% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dirs.R
\name{setup_dirs}
\alias{setup_dirs}
\title{Setup Required Directories}
\usage{
setup_dirs(site_dir)
}
\arguments{
\item{site_dir}{The root directory under which to create the new directories.}
}
\value{
Nothing, but creates directories.
}
\description{
This function creates the required directories if they do not already exist.
}
\examples{
setup_dirs('D:/Agisoft/ON/Block18_N_MX')
}
