% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_exif_data.R
\name{update_exif_data}
\alias{update_exif_data}
\title{Update EXIF Data with Corrected and Original Values}
\usage{
update_exif_data(site_dir)
}
\arguments{
\item{site_dir}{The directory path where the site-specific data is stored.}
}
\value{
NULL. The function updates the EXIF data in place.
}
\description{
This function updates the EXIF data of image files with corrected values
while also preserving the original values as new attributes.
}
\examples{
\dontrun{
update_exif_data("C:/path/to/site_dir")
}
}
